+++ .changes/unreleased/Under the Hood-20231128-170732.yaml
@@ -1,0 +1,6 @@
+kind: Under the Hood
+body: Move dbt.semver to dbt.common.semver and update references.
+time: 2023-11-28T17:07:32.172421-08:00
+custom:
+  Author: versusfacit
+  Issue: "9039"
