+++ core/dbt/version.py
@@ -1,14 +1,13 @@
+import glob
 import importlib
 import importlib.util
-import os
-import glob
 import json
-from typing import Iterator, List, Optional, Tuple
-
+import os
 import requests
 
-import dbt.exceptions
-import dbt.semver
+from typing import Iterator, List, Optional, Tuple
+
+import dbt.common.semver as semver
 
 from dbt.common.ui import green, red, yellow
 
@@ -34,13 +33,13 @@ def get_version_information() -> str:
     return "\n\n".join(msg_lines)
 
 
-def get_installed_version() -> dbt.semver.VersionSpecifier:
-    return dbt.semver.VersionSpecifier.from_version_string(__version__)
+def get_installed_version() -> semver.VersionSpecifier:
+    return semver.VersionSpecifier.from_version_string(__version__)
 
 
 def get_latest_version(
     version_url: str = PYPI_VERSION_URL,
-) -> Optional[dbt.semver.VersionSpecifier]:
+) -> Optional[semver.VersionSpecifier]:
     try:
         resp = requests.get(version_url, timeout=1)
         data = resp.json()
@@ -48,7 +47,7 @@ def get_latest_version(
     except (json.JSONDecodeError, KeyError, requests.RequestException):
         return None
 
-    return dbt.semver.VersionSpecifier.from_version_string(version_string)
+    return semver.VersionSpecifier.from_version_string(version_string)
 
 
 def _get_core_msg_lines(installed, latest) -> Tuple[List[List[str]], str]:
@@ -96,7 +95,7 @@ def _format_core_msg(lines: List[List[str]]) -> str:
     return msg + "\n".join(msg_lines)
 
 
-def _get_plugins_msg(installed: dbt.semver.VersionSpecifier) -> str:
+def _get_plugins_msg(installed: semver.VersionSpecifier) -> str:
     msg_lines = ["Plugins:"]
 
     plugins = []
@@ -122,9 +121,9 @@ def _get_plugins_msg(installed: dbt.semver.VersionSpecifier) -> str:
 
 
 def _get_plugin_msg_info(
-    name: str, version_s: str, core: dbt.semver.VersionSpecifier
+    name: str, version_s: str, core: semver.VersionSpecifier
 ) -> Tuple[str, bool]:
-    plugin = dbt.semver.VersionSpecifier.from_version_string(version_s)
+    plugin = semver.VersionSpecifier.from_version_string(version_s)
     latest_plugin = get_latest_version(version_url=get_package_pypi_url(name))
 
     needs_update = False
@@ -169,14 +168,12 @@ def _pad_lines(lines: List[List[str]], seperator: str = "") -> List[List[str]]:
 
     result: List[List[str]] = []
     for i, line in enumerate(lines):
-
         # add another list to hold padded strings
         if len(result) == i:
             result.append([""] * len(line))
 
         # iterate over columns in the line
         for j, item in enumerate(line):
-
             # the last column does not need padding
             if j == len(line) - 1:
                 result[i][j] = item
