+++ tests/unit/test_deps.py
@@ -20,7 +20,7 @@ from dbt.contracts.project import (
 )
 from dbt.config.project import PartialProject
 from dbt.contracts.project import PackageConfig
-from dbt.semver import VersionSpecifier
+from dbt.common.semver import VersionSpecifier
 from dbt.version import get_installed_version
 from dbt.common.dataclass_schema import ValidationError
 from dbt.flags import set_from_args
@@ -812,7 +812,6 @@ class TestPackageSpec(unittest.TestCase):
         self.assertEqual(resolved[0].version, "0.1.4a1")
 
     def test_validation_error_when_version_is_missing_from_package_config(self):
-
         packages_data = {"packages": [{"package": "dbt-labs-test/b", "version": None}]}
 
         with self.assertRaises(ValidationError) as exc:
@@ -822,7 +821,6 @@ class TestPackageSpec(unittest.TestCase):
         assert msg in str(exc.exception)
 
     def test_validation_error_when_namespace_is_missing_from_package_config(self):
-
         packages_data = {"packages": [{"package": "dbt-labs", "version": "1.0.0"}]}
 
         with self.assertRaises(ValidationError) as exc:
